jobs/ofac-scan/requirements.txt
requests==2.28.2
pygithub==1.55